loadlamb/requirements.txt
beautifulsoup4>=4.6.0
base58>=1.0.0
boto3>=1.7.37
click>=6.7
docb>=0.15.0
envs>=1.2.6
jinja2>=2.10
pip>=10
python-slugify>=1.2.5
PyYAML>=3.12
requests>=2.19.0
sammy>=0.3.10
troposphere>=2.3.0
unipath>=1.1
